Move in-memory server into rerun subcommand (#10891)
This PR moves over `rerun_server` to be a subcommand within the main
`rerun` binary. No other changes are included.

You can run it with this command:

```shell
pixi run rerun server -d /path/to/dataset
```

But you need to enable the `oss_feature` flag first.


## TODO
* [x] `@rerun-bot full-check`

---------
