+++ crates/store/re_server/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    re_build_tools::export_build_info_vars_for_crate("re_server");
+}
