+++ crates/store/re_server/src/lib.rs
@@ -6,7 +6,7 @@ mod server;
 mod store;
 
 pub use self::{
-    entrypoint::run,
+    entrypoint::Args,
     frontend::{FrontendHandler, FrontendHandlerBuilder, FrontendHandlerSettings},
     server::{Server, ServerBuilder, ServerError, ServerHandle},
 };
