+++ crates/top/rerun/Cargo.toml
@@ -35,6 +35,9 @@ default = [
   "server",
 ]
 
+## Support the `rerun auth` command
+auth = ["dep:re_viewer", "re_auth/cli"]
+
 ## Enable anonymized telemetry using our analytics SDK.
 analytics = [
   "dep:re_analytics",
@@ -52,25 +55,21 @@ clap = ["dep:clap"]
 ## for more information.
 data_loaders = ["dep:re_mcap", "re_sdk?/data_loaders"]
 
+## Access to Rerun's dataframe API and related types.
+dataframe = ["dep:re_dataframe"]
+
 ## Demo helpers for examples.
 demo = []
 
-## Access to Rerun's dataframe API and related types.
-dataframe = ["dep:re_dataframe"]
+## Integration with the [`ecolor`](https://crates.io/crates/ecolor/) crate.
+ecolor = ["re_types?/ecolor"]
 
 ## Add support for some math operations using [`glam`](https://crates.io/crates/glam/).
 ## Only relevant if feature `sdk` is enabled.
 glam = ["re_types?/glam"]
 
-## Add support for math type conversions using [`mint`](https://crates.io/crates/mint/).
-## Only relevant if feature `sdk` is enabled.
-mint = ["re_types?/mint"]
-
 ## Integration with the [`image`](https://crates.io/crates/image/) crate, plus JPEG support.
 image = ["re_types?/image"]
-
-## Integration with the [`ecolor`](https://crates.io/crates/ecolor/) crate.
-ecolor = ["re_types?/ecolor"]
 
 ## Integration with the [`log`](https://crates.io/crates/log/) crate.
 log = ["dep:env_filter", "dep:log"]
@@ -79,6 +78,10 @@ log = ["dep:env_filter", "dep:log"]
 ## This adds a lot of extra dependencies.
 map_view = ["re_viewer?/map_view"]
 
+## Add support for math type conversions using [`mint`](https://crates.io/crates/mint/).
+## Only relevant if feature `sdk` is enabled.
+mint = ["re_types?/mint"]
+
 ## Enable faster native video decoding with assembly.
 ## You need to install [nasm](https://github.com/netwide-assembler/nasm) to compile with this feature.
 nasm = ["re_video/nasm"]
@@ -86,6 +89,9 @@ nasm = ["re_video/nasm"]
 ## Support spawning a native viewer and allow to extend the viewer.
 ## This adds a lot of extra dependencies, so only enable this feature if you need it!
 native_viewer = ["dep:re_viewer", "dep:re_crash_handler"]
+
+## Enable the in-memory Rerun Server, useful for testing.
+oss_server = ["dep:re_server"]
 
 ## Enables integration with `re_perf_telemetry` (Tracy, Jaeger).
 ##
@@ -110,11 +116,11 @@ run = [
   "auth",
 ]
 
-## Support for running a gRPC server that listens to incoming log messages from a Rerun SDK.
-server = ["dep:re_grpc_server", "re_sdk/server", "tokio/signal"]
-
 ## Embed the Rerun SDK & built-in types and re-export all of their public symbols.
 sdk = ["dep:re_sdk", "dep:re_types"]
+
+## Support for running a gRPC server that listens to incoming log messages from a Rerun SDK.
+server = ["dep:re_grpc_server", "re_sdk/server", "tokio/signal"]
 
 ## Support serving a web viewer over HTTP.
 ##
@@ -127,7 +133,6 @@ sdk = ["dep:re_sdk", "dep:re_types"]
 ## to the compile time since it requires compiling and bundling the viewer as wasm.
 web_viewer = ["server", "dep:re_web_viewer_server", "re_sdk?/web_viewer"]
 
-auth = ["dep:re_viewer", "re_auth/cli"]
 
 [dependencies]
 re_build_info.workspace = true
@@ -173,6 +178,7 @@ re_global_context = { workspace = true, optional = true }
 re_grpc_server = { workspace = true, optional = true }
 re_mcap = { workspace = true, optional = true }
 re_sdk = { workspace = true, optional = true }
+re_server = { workspace = true, optional = true }
 re_types = { workspace = true, optional = true }
 re_viewer = { workspace = true, optional = true }
 re_web_viewer_server = { workspace = true, optional = true }
