+++ docs/content/reference/cli.md
@@ -17,11 +17,12 @@ The Rerun command-line interface:
 **Commands**
 
 * `analytics`: Configure the behavior of our analytics.
-* `mcap`: Manipulate the contents of .mcap files.
-* `rrd`: Manipulate the contents of .rrd and .rbl files.
-* `reset`: Reset the memory of the Rerun Viewer.
-* `man`: Generates the Rerun CLI manual (markdown).
 * `auth`: Authentication with the redap.
+* `man`: Generates the Rerun CLI manual (markdown).
+* `mcap`: Manipulate the contents of .mcap files.
+* `reset`: Reset the memory of the Rerun Viewer.
+* `rrd`: Manipulate the contents of .rrd and .rbl files.
+* `server`: In-memory Rerun data server.
 
 **Arguments**
 
@@ -232,6 +233,45 @@ Associate an email address with the current user.
 **Arguments**
 
 * `<EMAIL>`
+
+## rerun auth
+
+Authentication with the redap.
+
+**Usage**: `rerun auth <COMMAND>`
+
+**Commands**
+
+* `login`: Log into Rerun.
+* `token`: Retrieve the stored access token.
+
+## rerun auth login
+
+Log into Rerun.
+
+This command opens a page in your default browser, allowing you to log in to the Rerun data platform.
+
+Once you've logged in, your credentials are stored on your machine.
+
+To sign up, contact us through the form linked at <https://rerun.io/#open-source-vs-commercial>.
+
+**Usage**: `rerun auth login [OPTIONS]`
+
+**Options**
+
+* `--login-url <LOGIN_URL>`
+>
+> [Default: `https://rerun.io/login`]
+
+* `--no-open-browser <NO_OPEN_BROWSER>`
+> Post a link instead of directly opening in the browser.
+>
+> [Default: `false`]
+
+* `--force <FORCE>`
+> Trigger the full login flow even if valid credentials already exist.
+>
+> [Default: `false`]
 
 ## rerun mcap
 
@@ -579,41 +619,26 @@ Can be used to ensure that the current Rerun version can load the data.
 * `<PATH_TO_INPUT_RRDS>`
 > Paths to read from. Reads from standard input if none are specified.
 
-## rerun auth
+## rerun server
 
-Authentication with the redap.
+In-memory Rerun data server.
 
-**Usage**: `rerun auth <COMMAND>`
-
-**Commands**
-
-* `login`: Log into Rerun.
-* `token`: Retrieve the stored access token.
-
-## rerun auth login
-
-Log into Rerun.
-
-This command opens a page in your default browser, allowing you to log in to the Rerun data platform.
-
-Once you've logged in, your credentials are stored on your machine.
-
-To sign up, contact us through the form linked at <https://rerun.io/#open-source-vs-commercial>.
-
-**Usage**: `rerun auth login [OPTIONS]`
+**Usage**: `rerun server [OPTIONS]`
 
 **Options**
 
-* `--login-url <LOGIN_URL>`
+* `--addr <ADDR>`
+> Address to listen on.
 >
-> [Default: `https://rerun.io/login`]
+> [Default: `0.0.0.0`]
 
-* `--no-open-browser <NO_OPEN_BROWSER>`
-> Post a link instead of directly opening in the browser.
+* `-p, --port <PORT>`
+> Port to bind to.
 >
-> [Default: `false`]
+> [Default: `51234`]
 
-* `--force <FORCE>`
-> Trigger the full login flow even if valid credentials already exist.
->
-> [Default: `false`]
+* `-d, --dataset <DATASETS>`
+> Load a directory of RRD as dataset (can be specified multiple times).
+
+* `-V, --version `
+> Print version.
